Added Error Manager class, it handles the erros that come from the server and puts them in a nice format
